Admin: Fix various tests to keep parity with AWS
