+++ tests/test_stepfunctions/parser/test_stepfunctions_dynamodb_integration.py
@@ -269,29 +269,21 @@ def test_send_task_failure_invalid_token(table_name=None, sleep_time=0):
                     client.send_task_failure(taskToken="bad_token", error="test error")
 
                 # Verify
-                assert (
-                    exc.value.response["Error"]["Code"] == "UnrecognizedClientException"
-                )
+                assert exc.value.response["Error"]["Code"] == "InvalidToken"
                 assert (
                     exc.value.response["Error"]["Message"]
-                    == "The security token included "
-                    "in the request is invalid."
+                    == "Invalid Token: 'Invalid token'"
                 )
 
                 # Execute
                 with pytest.raises(ClientError) as exc:
-                    client.send_task_success(
-                        taskToken="bad_token", output="test output"
-                    )
+                    client.send_task_success(taskToken="bad_token", output="output")
 
                 # Verify
-                assert (
-                    exc.value.response["Error"]["Code"] == "UnrecognizedClientException"
-                )
+                assert exc.value.response["Error"]["Code"] == "InvalidToken"
                 assert (
                     exc.value.response["Error"]["Message"]
-                    == "The security token included "
-                    "in the request is invalid."
+                    == "Invalid Token: 'Invalid token'"
                 )
 
     verify_execution_result(
